requirements.txt
# numpy==1.26.4
# pandas==2.2.1
# matplotlib==3.8.3
# scikit-learn==1.4.1.post1
#

langchain==0.1.12


python-dotenv==1.0.1
streamlit==1.32.2
openai==1.14.1
# huggingface_hub==0.21.4
# pypdf2==3.0.1
# faiss-cpu==1.7.4
# tiktoken==0.6.0
# langchain-community==0.0.24